Clean expressions

+++ src/main/java/ch/njol/skript/expressions/ExprArithmetic.java
@@ -56,8 +56,8 @@ public class ExprArithmetic extends SimpleExpression<Number> {
 			@Override
 			public Number calculate(final Number n1, final Number n2, final boolean integer) {
 				if (integer)
-					return Long.valueOf(n1.longValue() + n2.longValue());
-				return Double.valueOf(n1.doubleValue() + n2.doubleValue());
+					return n1.longValue() + n2.longValue();
+				return n1.doubleValue() + n2.doubleValue();
 			}
 		},
 		MINUS('-') {
@@ -65,8 +65,8 @@ public class ExprArithmetic extends SimpleExpression<Number> {
 			@Override
 			public Number calculate(final Number n1, final Number n2, final boolean integer) {
 				if (integer)
-					return Long.valueOf(n1.longValue() - n2.longValue());
-				return Double.valueOf(n1.doubleValue() - n2.doubleValue());
+					return n1.longValue() - n2.longValue();
+				return n1.doubleValue() - n2.doubleValue();
 			}
 		},
 		MULT('*') {
@@ -74,8 +74,8 @@ public class ExprArithmetic extends SimpleExpression<Number> {
 			@Override
 			public Number calculate(final Number n1, final Number n2, final boolean integer) {
 				if (integer)
-					return Long.valueOf(n1.longValue() * n2.longValue());
-				return Double.valueOf(n1.doubleValue() * n2.doubleValue());
+					return n1.longValue() * n2.longValue();
+				return n1.doubleValue() * n2.doubleValue();
 			}
 		},
 		DIV('/') {
@@ -86,9 +86,9 @@ public class ExprArithmetic extends SimpleExpression<Number> {
 					final long div = n2.longValue();
 					if (div == 0)
 						return Long.MAX_VALUE;
-					return Long.valueOf(n1.longValue() / div);
+					return n1.longValue() / div;
 				}
-				return Double.valueOf(n1.doubleValue() / n2.doubleValue());
+				return n1.doubleValue() / n2.doubleValue();
 			}
 		},
 		EXP('^') {
@@ -96,8 +96,8 @@ public class ExprArithmetic extends SimpleExpression<Number> {
 			@Override
 			public Number calculate(final Number n1, final Number n2, final boolean integer) {
 				if (integer)
-					return Long.valueOf((long) Math.pow(n1.longValue(), n2.longValue()));
-				return Double.valueOf(Math.pow(n1.doubleValue(), n2.doubleValue()));
+					return (long) Math.pow(n1.longValue(), n2.longValue());
+				return Math.pow(n1.doubleValue(), n2.doubleValue());
 			}
 		};
